Check if DB upgrade has already been performed
For those websites where things don’t behave as expected

+++ database/schema/420.php
@@ -13,14 +13,56 @@ class Red_Database_420 extends Red_Database_Upgrader {
 	}
 
 	protected function remove_module_id( $wpdb ) {
+		if ( ! $this->has_module_id( $wpdb ) ) {
+			return true;
+		}
+
 		return $this->do_query( $wpdb, "ALTER TABLE `{$wpdb->prefix}redirection_logs` DROP `module_id`" );
 	}
 
 	protected function remove_group_id( $wpdb ) {
+		if ( ! $this->has_module_id( $wpdb ) ) {
+			return true;
+		}
+
 		return $this->do_query( $wpdb, "ALTER TABLE `{$wpdb->prefix}redirection_logs` DROP `group_id`" );
 	}
 
+	private function has_module_id( $wpdb ) {
+		$existing = $wpdb->get_row( "SHOW CREATE TABLE `{$wpdb->prefix}redirection_logs`", ARRAY_N );
+
+		if ( isset( $existing[1] ) && strpos( strtolower( $existing[1] ), 'module_id' ) !== false ) {
+			return true;
+		}
+
+		return false;
+	}
+
+	private function has_log_domain( $wpdb ) {
+		$existing = $wpdb->get_row( "SHOW CREATE TABLE `{$wpdb->prefix}redirection_logs`", ARRAY_N );
+
+		if ( isset( $existing[1] ) && strpos( strtolower( $existing[1] ), 'domain` varchar' ) !== false ) {
+			return true;
+		}
+
+		return false;
+	}
+
+	private function has_404_domain( $wpdb ) {
+		$existing = $wpdb->get_row( "SHOW CREATE TABLE `{$wpdb->prefix}redirection_404`", ARRAY_N );
+
+		if ( isset( $existing[1] ) && strpos( strtolower( $existing[1] ), 'domain` varchar' ) !== false ) {
+			return true;
+		}
+
+		return false;
+	}
+
 	protected function add_extra_logging( $wpdb ) {
+		if ( $this->has_log_domain( $wpdb ) ) {
+			return true;
+		}
+
 		// Update any URL with a double slash at the end
 		$this->do_query( $wpdb, "ALTER TABLE `{$wpdb->prefix}redirection_logs` ADD `domain` VARCHAR(255) NULL DEFAULT NULL AFTER `url`" );
 		$this->do_query( $wpdb, "ALTER TABLE `{$wpdb->prefix}redirection_logs` ADD `http_code` INT(11) unsigned NOT NULL DEFAULT 0 AFTER `referrer`" );
@@ -31,10 +73,18 @@ class Red_Database_420 extends Red_Database_Upgrader {
 	}
 
 	protected function allow_null_agent( $wpdb ) {
+		if ( $this->has_log_domain( $wpdb ) ) {
+			return true;
+		}
+
 		return $this->do_query( $wpdb, "ALTER TABLE `{$wpdb->prefix}redirection_logs` CHANGE COLUMN `agent` `agent` MEDIUMTEXT NULL" );
 	}
 
 	protected function add_extra_404( $wpdb ) {
+		if ( $this->has_404_domain( $wpdb ) ) {
+			return true;
+		}
+
 		// Update any URL with a double slash at the end
 		$this->do_query( $wpdb, "ALTER TABLE `{$wpdb->prefix}redirection_404` ADD `domain` VARCHAR(255) NULL DEFAULT NULL AFTER `url`" );
 		$this->do_query( $wpdb, "ALTER TABLE `{$wpdb->prefix}redirection_404` ADD `http_code` INT(11) unsigned NOT NULL DEFAULT 0 AFTER `referrer`" );
@@ -44,6 +94,10 @@ class Red_Database_420 extends Red_Database_Upgrader {
 	}
 
 	protected function convert_404_url_to_text( $wpdb ) {
+		if ( $this->has_404_domain( $wpdb ) ) {
+			return true;
+		}
+
 		// Same as log table
 		$this->do_query( $wpdb, "ALTER TABLE `{$wpdb->prefix}redirection_404` DROP INDEX `url`" );
 		return $this->do_query( $wpdb, "ALTER TABLE `{$wpdb->prefix}redirection_404` CHANGE COLUMN `url` `url` MEDIUMTEXT NOT NULL" );
